fix: ensure newline at EOF in READMEs
