+++ README.es.md
@@ -58,7 +58,7 @@ python3.10 -m venv .venv
 source .venv/bin/activate
 pip install -r requirements.txt
 export OPENAI_API_KEY=tu-clave
-python main.py --readme README.md --improved README.improved.md
+python run_improver.py --readme README.md --improved README.improved.md
 ```
 
 Opciones principales:
@@ -88,7 +88,7 @@ jobs:
           python-version: "3.10"
       - run: pip install "openai>=1.0" python-dotenv markdown2
       - run: echo "OPENAI_API_KEY=${{ secrets.OPENAI_API_KEY }}" >> $GITHUB_ENV
-      - run: python main.py
+      - run: python run_improver.py
       - name: Replace README with improved version
         run: |
           mv README.improved.md README.md
